Disallow `#[schemars(regex(path = ...))]` and  `#[validate(regex(pattern = ...))]`
For compatibility with the validator crate, when the `pattern` form is given a string literal, that string's contents will be parsed as an expression. This is due to the behaviour of the darling crate (https://github.com/TedDriggs/darling/issues/229). I'm concerned that this behaviour may change in a future version of darling/validator, so to minimise its exposure in schemars's public API, I'm restricting its usage now.